Information can now be stored in a RAG for the LLMCharacter to check.
WIP in Niko Chatbot.
New GameObjects: RAG, RAGData.

RAG uses a reference to the LLM in the scene.

RAGData uses the RAG and a text document to store the information in. This is then used in the NewChatBot.cs before the prompt is sent to the LLM.

Currently the idea is that RAGData is single compontent and the individual characters will each have a seperate Text Asset on them that will serve as the referance. Either that or multiple RAGs and RAGDatas are attached to each character. Depends on how it performs.